% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.saegamlss_class.R
\name{as.data.frame.saegamlss}
\alias{as.data.frame.saegamlss}
\title{as.data.frame}
\usage{
\method{as.data.frame}{saegamlss}(x, row.names = NULL, optional = FALSE, n = 1, ...)
}
\arguments{
\item{x}{any R object of class "saegamlss"}

\item{row.names}{NULL (Default) or a character vector giving the row names for the data frame. Missing values are not allowed}

\item{optional}{logical. If TRUE, setting row names and converting column names (to syntactic names: see make.names) is optional}

\item{n}{The population to be transformed in a data.frame. To be defined only for data_gen(). Default is 1}

\item{...}{Other parameters}
}
\value{
A dataset for each population that has been generated
}
\description{
transforms in data.frame the x of class "saegamlss".
}
\examples{

data_p = data_gen(
  Ni = rep(10, 4), D = 4, ty = "no", k = 100, b1 = 4,
  x1 = rnorm(40, 0, 1), b2 = NULL, x2 = NULL, b3 = NULL,
  x3 = NULL, b4 = NULL, x4 = NULL, xh = NULL,
  Dis=rNO, l = c(identity), sigma = 6, sigmah = NULL,
  sigmae = 22, costh = NULL, M = 2
  )

as.data.frame(x = data_p)
}
\seealso{
[data_gen()]
}
