% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atkinson_Logno.R
\name{atkinson.logno}
\alias{atkinson.logno}
\title{Atkinson index Log-Normal distribution}
\usage{
atkinson.logno(sigma, epsilon = NULL)
}
\arguments{
\item{sigma}{The estimated value of sigma}

\item{epsilon}{The value for the poverty aversion parameter. Default value is set to 1.}
}
\value{
The Atkinson index
}
\description{
Compute the Atkinson index with Log-Normal distribution
}
