% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_saegamlss_class.R
\name{coef.saegamlss}
\alias{coef.saegamlss}
\title{Estimated regression parameters}
\usage{
\method{coef}{saegamlss}(object, parameters = c("mu", "sigma", "nu", "tau"), ...)
}
\arguments{
\item{object}{An object of class saegamlss}

\item{parameters}{The parameter(S) for which the Estimated regression parameters has to be extracted. Default is c("mu", "sigma", "nu", "tau")}

\item{...}{Other parameters}
}
\value{
A dataset with the estimated Estimated regression parameters
}
\description{
Extract Estimated regression parameters for object of class "saegamlss"
}
\note{
The function ranef is firstly imported from nlme and than adapted for object of class "saegamlss"
}
\examples{
index_est <- sa_p_index(sample = s_data, y = "y",
                        sa = "sa", fdis = "LOGNO",
                        sigma.f = TRUE, index = "all")

coef(object = index_est, parameters = c("mu", "sigma"))

est <- est_saegamlss(
  sample = s_data, nonsample = pop_data, y_dip="y",
  sa="sa", f1 = y ~ x1 + random(sa), f2 = y ~ x2 + random(sa),
  f3 = NULL, f4 = NULL, fdis = NO, R = 2,
  Dis = rNO, param = "both",
  tau.fix = NULL, nu.fix = NULL
)

coef(object = est, parameters = c("mu", "sigma"))

}
