% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{data_gen}
\alias{data_gen}
\title{Data generation}
\usage{
data_gen(
  Ni,
  D,
  k,
  b1,
  x1,
  b2 = NULL,
  x2 = NULL,
  b3 = NULL,
  x3 = NULL,
  b4 = NULL,
  x4 = NULL,
  xh = NULL,
  Dis,
  l,
  sigma = NULL,
  sigmah = NULL,
  sigmae = NULL,
  ty = NULL,
  costh = NULL,
  id = NULL,
  M = NULL,
  seed = 123
)
}
\arguments{
\item{Ni}{1xD vector containing the number of units (in population) for each area}

\item{D}{Number of areas}

\item{k}{A vector, maximum length equal to 4, of constants (intercepts) used to generate data, i.e. mu=k+b1*x}

\item{b1}{A vector of regression coefficients used for mu}

\item{x1}{A matrix/vector or a dataset where each column is a covariate used for mu}

\item{b2}{A vector of regression coefficients used for sigma}

\item{x2}{A matrix/vector or a dataset where each column is a covariate used for sigma}

\item{b3}{A vector of regression coefficients used for nu}

\item{x3}{A matrix/vector or a dataset where each column is a covariate used for nu}

\item{b4}{A vector of regression coefficients used for tau}

\item{x4}{A matrix/vector or a dataset where each column is a covariate used for tau}

\item{xh}{A vector of covariate used to generate heteroschedasticity}

\item{Dis}{Type of distribution in form of rDis where Dis is one of the distribution allowed by GAMLSS.}

\item{l}{A vector,  of maximum length equal to 4, in which each space is a function (the inverse of the link-function) to be applied to the corresponding parameter, i.e. mu, sigma, nu and tau}

\item{sigma}{A vector,  of maximum length equal to 4, of standard deviation used to generate random effects}

\item{sigmah}{The standard deviation used to generate heteroschedasticity}

\item{sigmae}{The standard deviation used to generate the error term when the link-function is the identity one}

\item{ty}{To be specified equal to "no" if the generation process is mu=k+b1*x+e  or equal to "het" if the generation process have to include heteroschedasticity}

\item{costh}{A constant used to generate heteroschedasticity}

\item{id}{Unit id}

\item{M}{Number of replicates. Default is 1}

\item{seed}{The seed. Default is 123}
}
\value{
An object of class "saegamlss" of length M. Each element of this list is a data frame that has: the dependent variable y, a set of fixed covariates, the small area column (sa) and the id column (id)
}
\description{
A tool to generate a population with a dependent variable defined in terms of a given distribution and a set of (fixed) covariates.
The dependent variable can be generated for more than 100 distributions, each with a maximum of 4 parameters. Each parameter can be defined in terms of covariates and random effects.
When the link function is the identity function, the dependent variable may include an additional error term. For the normal distribution, it is possible to add heteroscedasticity.
}
\note{
The definition of the heteroschedastic term follows Ramirez-Aldana, R., & Naranjo, L. (2021).
}
\examples{
# Normal data (2 populations)

data_gen(
  Ni = rep(10, 4), D = 4, ty = "no", k = 100, b1 = 4,
  x1 = rnorm(40, 0, 1), b2 = NULL, x2 = NULL, b3 = NULL,
  x3 = NULL, b4 = NULL, x4 = NULL, xh = NULL,
  Dis=rNO, l = c(identity), sigma = 6, sigmah = NULL,
  sigmae = 22, costh = NULL, M = 2
)

# Heteroschedastic Normal data

data_gen(
  Ni = rep(10, 4), D = 4, ty = "het", k = 100, b1 = 4,
  x1 = rnorm(40, 0, 1), b2 = NULL, x2 = NULL, b3 = NULL,
  x3 = NULL, b4 = NULL, x4 = NULL, xh = rnorm(40, 0, 1),
  Dis=rNO, l = c(identity), sigma = 6, sigmah = 1, sigmae = 22,
  costh = 0.1
)

}
\references{
Mori, L., & Ferrante, M. R. (2024). Small area estimation under unit-level generalized additive models for location, scale and shape. Journal of Survey Statistics and Methodology, 1–37
 Ramirez-Aldana, R., & Naranjo, L. (2021). Random intercept and linear mixed models including heteroscedasticity in a logarithmic scale: Correction terms and prediction in the original scale. PloS one, 16(4), e0249910.
 Rigby, R. A., & Stasinopoulos, D. M. (2005). Generalized additive models for location, scale and shape. Journal of the Royal Statistical Society: Series C (Applied Statistics), 54(3), 507–554.
}
\author{
Lorenzo Mori and Maria Rosaria Ferrante
}
