% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_HT.R
\name{direct_HT}
\alias{direct_HT}
\title{Direct Horvitz-Thompson estimator of the mean}
\usage{
direct_HT(
  sample,
  y,
  sa,
  pi,
  pi2 = NULL,
  N = NULL,
  var_est = TRUE,
  var_method = "bootstrapSRS",
  B = 100,
  fpc = TRUE
)
}
\arguments{
\item{sample}{The data-set of sampled units containing y, sa, and pi}

\item{y}{The name of the  variable of interest}

\item{sa}{The name of the Small Areas}

\item{pi}{The name of the first order inclusion probability}

\item{pi2}{A square matrix of the joint inclusion probabilities. Needed for the "LinHT" variance estimator}

\item{N}{A numeric value of the population size for each area. If NULL, it is estimated with the sum of the inverse of the pis}

\item{var_est}{A logical indicating whether or not to compute a variance estimator. Default is TRUE}

\item{var_method}{The method to use when computing the variance estimator. Options are a Taylor linearized technique: "LinHB"= Hajek-Berger estimator, "LinHH" = Hansen-Hurwitz estimator, "LinHTSRS" = Horvitz-Thompson estimator under simple random sampling without replacement, and "LinHT" = Horvitz-Thompson estimator or a resampling technique: "bootstrapSRS" = bootstrap variance estimator under simple random sampling without replacement. The default is "bootstrapSRS"}

\item{B}{The number of bootstrap samples if computing the bootstrap variance estimator. Default is 100}

\item{fpc}{Default to TRUE, logical for whether or not the variance calculation should include a finite population correction when calculating the "LinHTSRS" or the "SRSbootstrap" variance estimator}
}
\value{
An object of class "saegamlss_class" with the Horvitz-Thompson estimate of the mean, the variance, the Standard Deviation and the Coefficient of Variation for each Area
}
\description{
Compute the Horvitz-Thompson estimator for a finite population mean or proportion based on sample data collected from a complex sampling design for small areas.
}
\examples{

##################
###Using s_data###
##################

#Using the default bootstrap

dir = direct_HT(y = "y", sa = "sa", pi = "pi", sample = s_data)

dir$results_HT

#Using HH linearization
dir = direct_HT(y = "y", sa = "sa", pi = "pi", sample = s_data,
                var_method = "LinHH")

dir$results_HT

}
\references{
Wolter, Kirk M., and Kirk M. Wolter. Introduction to variance estimation. Vol. 53. New York: Springer, 2007.
}
\author{
Lorenzo Mori and Maria Rosaria Ferrante
}
