% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_saegamlss.R
\name{est_saegamlss}
\alias{est_saegamlss}
\title{Monte Carlo estimation of mean and HCR based on SAE GAMLSS}
\usage{
est_saegamlss(
  sample,
  nonsample,
  D,
  Ni,
  ni,
  f1,
  f2 = NULL,
  f3 = NULL,
  f4 = NULL,
  fdis,
  nRS = NULL,
  nCG = NULL,
  R = NULL,
  Dis,
  np,
  param = NULL,
  seed = NULL,
  tau.fix = NULL,
  nu.fix = NULL,
  z = NULL
)
}
\arguments{
\item{sample}{A dataset of sampled units. With dependent variable (named y), covariates and small area (named sa)}

\item{nonsample}{A dataset of non-sampled units. With covariates and small area (named sa)}

\item{D}{Number of Areas}

\item{Ni}{1xD vector containing the number of units (in population) for each area}

\item{ni}{1xD vector containing the number of sampled-units for each area}

\item{f1}{A formula object for fitting a model to the mu parameter, e.g.f1=y~x+random(sa)}

\item{f2}{A formula object for fitting a model to the sigma parameter, e.g.f2=y~x+random(sa)}

\item{f3}{A formula object for fitting a model to the nu parameter, e.g.f3=y~x+random(sa)}

\item{f4}{A formula object for fitting a model to the tau parameter, e.g.f4=y~x+random(sa)}

\item{fdis}{The distribution family of the GAMLSS object}

\item{nRS}{Number of loop to do with the RS() algorithm}

\item{nCG}{Number of loop to do with the CG() algorithm}

\item{R}{Number of Monte-Carlo repetition. Default is 50}

\item{Dis}{Type of distribution in form of rDis where Dis is one of the distribution allowed by GAMLSS}

\item{np}{Number of parameters of the distribution. i.e. for the normal distribution np=2, for the GB2 distribution np=4}

\item{param}{The parameter to estimate, "mean" or "HCR". Default is both.}

\item{seed}{The seed}

\item{tau.fix}{A value to be fixed to 1 to obtain the Singh-Maddala distribution. Dist must be GB2 and np=4}

\item{nu.fix}{A value to be fixed to 1 to obtain the Dagum distribution. Dist must be GB2 and np=4}

\item{z}{The Poverty line. Default is equal to 0.6 of the dependent variable}
}
\value{
a list of two list. The first list, named est has:
 ME 1xD vector with the estimate of the mean and
 HCR 1xD vector with the estimate of the HCR.
 The second, named input_var has:
 fit a GAMLSS object with all the components of the regression
 f1 a formula object used for fitting a model to the mu parameter
 f2 a formula object used for fitting a model to the sigma parameter
 f3 a  formula object used for fitting a model to the nu parameter
 f4 a formula object used for fitting a model to the tau parameter
 np the number of the parameters of the chosen distribution
 nRS the number of loop used within the RS() algorithm. Default is 150.
 nCG the number of loop used within the CG() algorithm. Default is 150.
 R the number of loop used to obtain the monte-carlo estimation
 D the number of area
 Ni 1xD vector containing the number of units (in population) for each area
 ni 1xD vector containing the number of sampled-units for each area
 originaldata The dataset of sampled-units
 fids The distribution used for the estimation
 z The poverty line
}
\description{
A function to estimate the mean or/and the HCR     derived from unit-level small area estimation based on generalized additive models for location, scale and shape
}
\note{
With "object"$input_var$fit is possible to use all the classical function used by gamlss
 The Small Area have to be denoted with a number from 1 to D
}
\examples{
# Generate data
#
dep.y <- data_gen(
  Ni = rep(10, 4), D = 4, M = 2, ty = "no", k = 1, b1 = 10,
  x1 = rnorm(40, 0, 1), b2 = NULL, x2 = NULL, b3 = NULL,
  b4 = NULL, x4 = NULL, xh = NULL, Dis = NO,
  l = c(identity), sigma = 6, sigmah = NULL,
  sigmae = 2, costh = NULL, seed = 1234
)

data <- dep.y[[1]]
#
# sample data with a sample fraction of 0.1
#
library(splitstackshape)
# sample data
#
sample <- stratified(data, "sa", size = 0.1)
# nonsample data
#
nonsample <- nonsample(data = data, sample = sample, id = id)
# estimate
est_saegamlss(
  sample = sample, nonsample = nonsample,
  D = 4, Ni = rep(10, 4), ni = rep(1, 4),
  f1 = y ~ x1 + random(sa), f2 = NULL, f3 = NULL,
  f4 = NULL, fdis = NO, nRS = 150, nCG = 150, R = 200,
  Dis = rNO, np = 2, param = NULL,
  seed = 1234, tau.fix = NULL, nu.fix = NULL
)
}
\references{
Mori, L., & Ferrante, M. R. (2023). Small area estimation under unit-level generalized additive models for location, scale and shape. arXiv e-prints, arXiv-2302.
 Graf, M., Marin, J. M., & Molina, I. (2019). A generalized mixed model for skewed distributions applied to small area estimation. Test, 28(2), 565–597.
}
\author{
Lorenzo Mori and Maria Rosaria Ferrante
}
