% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gini_Gamma.R
\name{gini.gamma}
\alias{gini.gamma}
\title{Gini index Gamma distribution}
\usage{
gini.gamma(sigma)
}
\arguments{
\item{sigma}{The estimated value of sigma}
}
\value{
The Gini index
}
\description{
Compute the Gini index with Gamma distribution
}
