% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gini_Logno.R
\name{gini.logno}
\alias{gini.logno}
\title{Gini index Log-Normal distribution}
\usage{
gini.logno(sigma)
}
\arguments{
\item{sigma}{The estimated value of sigma}
}
\value{
The Gini index
}
\description{
Compute the Gini index with Log-Normal distribution
}
