% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gini_Pareto.R
\name{gini.pareto}
\alias{gini.pareto}
\title{Gini index Pareto distribution}
\usage{
gini.pareto(sigma)
}
\arguments{
\item{sigma}{The estimated value of sigma}
}
\value{
The Gini index
}
\description{
Compute the Gini index with Pareto distribution
}
