% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gini_Weibull.R
\name{gini.weibull}
\alias{gini.weibull}
\title{Gini index Weibull distribution}
\usage{
gini.weibull(sigma)
}
\arguments{
\item{sigma}{The estimated value of sigma}
}
\value{
The Gini index
}
\description{
Compute the Gini index with Weibull distribution
}
