% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_selection.R
\name{m_selection}
\alias{m_selection}
\title{Small area model selection within GAMLSS}
\usage{
m_selection(
  sample,
  y,
  f_cov,
  kp = 2,
  ndis = 3,
  R = 200,
  k = 7,
  type = NULL,
  fix_dis = NULL,
  supp = TRUE,
  seed = 123
)
}
\arguments{
\item{sample}{The dataset with sampled units}

\item{y}{The name of dependent variable}

\item{f_cov}{A formula containing all the possible covariates and/or additive terms (i.e. x1+x2+x3+random(x4))}

\item{kp}{the penalty for the GAIC (Step 1) with default values kp = 2 the standard AIC}

\item{ndis}{Number of distribution to be consider at the first step. Default is 3}

\item{R}{Number of loop to be done within the k-fold cross validation}

\item{k}{Number of fold in k-fold cross validation. Default is 7}

\item{type}{Type of step 2 to be done. A or B. Default is B (see references)}

\item{fix_dis}{A distribution to be tested even if is not selected within the ndis at step 1}

\item{supp}{Suppress the warning in the three step. Default is TRUE}

\item{seed}{The seed. Default is 123}
}
\value{
An object of class "saegamlss_class", named results, with the results of each steps and the GAIC of all the tested distributions at step 1.
}
\description{
The function selects the final SAE-GAMLSS model following three steps as described by Mori and Ferrante (2024).
}
\examples{


##################
###Using s_data###
##################


sol <- m_selection(sample = s_data, y = "y",
                   f_cov = ~x1+x2+x3+random(sa), ndis = 1,
                   R = 2, k = 2, type = "B", fix_dis = "NO",
                   seed = 123, kp = 2)

#Step 1

sol$step1

#Step 2

sol$step2

#Step 3

sol$step3

}
\references{
Mori, L., & Ferrante, M. R. (2024). Small area estimation under unit-level generalized additive models for location, scale and shape. Journal of Survey Statistics and Methodology, 1–37
 Rigby, R. A., & Stasinopoulos, D. M. (2005). Generalized additive models for location, scale and shape. Journal of the Royal Statistical Society: Series C (Applied Statistics), 54(3), 507–554.
}
\author{
Lorenzo Mori and Maria Rosaria Ferrante
}
