% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_2.R
\docType{data}
\name{pi_2}
\alias{pi_2}
\title{second order inclusion probability data}
\format{
An object of class \code{"data.frame"}
}
\source{
{Created by authors following the Household Budget Survey structure}
}
\usage{
data(pi_2)
}
\description{
second order inclusion probability data
}
\keyword{datasets}
