% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_saegamlss.R
\name{plot.saegamlss_class}
\alias{plot.saegamlss_class}
\title{Plot of object of class "saegamlss_class}
\usage{
\method{plot}{saegamlss_class}(
  obj,
  compare.Gini = NULL,
  compare.Theil = NULL,
  compare.Atkinson = NULL,
  compare.Mean = NULL,
  compare.HCR = NULL,
  ...
)
}
\arguments{
\item{obj}{An object of class "saegamlss_clas"}

\item{compare.Gini}{The MSE of a second estimator to be compared for the Gini index}

\item{compare.Theil}{The MSE of a second estimator to be compared for the Theil index}

\item{compare.Atkinson}{The MSE of a second estimator to be compared for the Atkinson index}

\item{compare.Mean}{The MSE of a second estimator to be compared for the Mean}

\item{compare.HCR}{The MSE of a second estimator to be compared for the HCR}

\item{...}{Additional parameters}
}
\value{
Return a plot for object of class "saegamlss_class"
}
\description{
Plot of object of class "saegamlss_class
}
\examples{

# est_saegamlss

 dep.y <- data_gen(
  Ni = rep(10, 4), D = 4, M = 2, ty = "no", k = 1, b1 = 10,
  x1 = rnorm(40, 0, 1), b2 = NULL, x2 = NULL, b3 = NULL,
  b4 = NULL, x4 = NULL, xh = NULL, Dis = NO,
  l = c(identity), sigma = 6, sigmah = NULL,
  sigmae = 2, costh = NULL, seed = 1234
)

data <- dep.y[[1]]
#
# sample data with a sample fraction of 0.1
#
library(splitstackshape)
# sample data
#
sample <- stratified(data, "sa", size = 0.1)
# nonsample data
#
nonsample <- nonsample(data = data, sample = sample, id = id)
# estimate
est <- est_saegamlss(
  sample = sample, nonsample = nonsample,
  D = 4, Ni = rep(10, 4), ni = rep(1, 4),
  f1 = y ~ x1 + random(sa), f2 = NULL, f3 = NULL,
  f4 = NULL, fdis = NO, nRS = 150, nCG = 150, R = 200,
  Dis = rNO, np = 2, param = NULL,
  seed = 1234, tau.fix = NULL, nu.fix = NULL
)
plot(est)

#mse_saegamlss

dep.y <- data_gen(
  Ni = rep(10, 4), D = 4, M = 2, ty = "no", k = 4, b1 = 100,
  x1 = rnorm(40, 0, 1), b2 = NULL, x2 = NULL, b3 = NULL,
  x3 = NULL, b4 = NULL, x4 = NULL, xh = NULL,
  Dis, l = c(identity), sigma = 6, sigmah = NULL,
  sigmae = 22, costh = NULL, seed = 1234
)
data <- dep.y[[1]]
#
# sample data with a sample fraction of 0.5
#
library(splitstackshape)
# sample data
#
sample <- stratified(data, "sa", size = 0.5)
# nonsample data
#
nonsample <- nonsample(data = data, sample = sample, id = id)
# estimate
est <- est_saegamlss(
  sample = sample, nonsample = nonsample, D = 4,
  Ni = rep(10, 4), ni = rep(1, 4), f1 = y ~ x1 + random(sa),
  f2 = NULL, f3 = NULL, f4 = NULL, fdis = NO, nRS = 150,
  nCG = 150, R = 2, Dis = rNO, np = 2, param = NULL,
  seed = 1234, tau.fix = NULL, nu.fix = NULL
)
#
# covariates
#
x <- data.frame(rep(1, nrow(data)), "x1" = data$x1)
#
# compute the MSE
#
MSE <- mse_saegamlss(
  est = est, D = 4, Ni = rep(10, 4), loop = 2,
  l = c(identity), Dis = rNO, Iden = TRUE,
  samplesize = 0.1, data = data, cov1 = x, cov2 = NULL, cov3 = NULL,
  cov4 = NULL, seed = 1234
)

plot(MSE, compare.Mean= runif(4,0, 2), compare.HCR= runif(4,2, 4))
}
