% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/povinc.R
\name{povinc}
\alias{povinc}
\title{Head Count Ratio}
\usage{
povinc(y, z = NULL)
}
\arguments{
\item{y}{A vector with numerical valeus. Usually income or consumption data.}

\item{z}{The poverty line. Default is 0.6*median(y)}
}
\value{
The Head Count Ratio
}
\description{
Compute the Head Count Ration
}
\examples{
y=runif(100, 0, 1000)
povinc(y)
}
\author{
Lorenzo Mori and Maria Rosaria Ferrante
}
