% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_saegamlss.R
\name{print.saegamlss_class}
\alias{print.saegamlss_class}
\title{Print of x of class "saegamlss_class"}
\usage{
\method{print}{saegamlss_class}(x, ...)
}
\arguments{
\item{x}{An x of class "saegamlss_class"}

\item{...}{Additional parameters}
}
\value{
The Print of x of class "saegamlss_class"
}
\description{
Print for an x of class "saegamlss_class"
}
\examples{

#est_saegamlss

dep.y <- data_gen(
  Ni = rep(10, 4), D = 4, M = 1, ty = "no", k = 1, b1 = 10,
  x1 = rnorm(40, 0, 1), b2 = NULL, x2 = NULL, b3 = NULL,
  b4 = NULL, x4 = NULL, xh = NULL, Dis = NO,
  l = c(identity), sigma = 6, sigmah = NULL,
  sigmae = 2, costh = NULL
)

print(dep.y)
}
