% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran_eff.R
\name{ran_eff}
\alias{ran_eff}
\title{Extract random effects for object of class "saegamlss_class"}
\usage{
ran_eff(model, parameters = c("mu", "sigma", "nu", "tau"))
}
\arguments{
\item{model}{An object of class saegamlss_class}

\item{parameters}{the distribution parameters from which random effects are extracted. Default is  c("mu", "sigma", "nu", "tau")}
}
\value{
A dataset with the estiamted random effects
}
\description{
Extract random effects for object of class "saegamlss_class"
}
\examples{

index_est <- sa_p_index(sample = s_data, y = "y",
                        sa = "sa", fdis = "LOGNO",
                        sigma.f = TRUE, index = "all")

ran_eff(model = index_est, parameters = c("mu", "sigma"))

}
