% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef_saegamlss.R
\name{ranef.saegamlss}
\alias{ranef.saegamlss}
\title{Random effects}
\usage{
\method{ranef}{saegamlss}(object, parameters = c("mu", "sigma", "nu", "tau"), ...)
}
\arguments{
\item{object}{An object of class saegamlss}

\item{parameters}{The parameter(S) for which the random effects has to be extracted. Default is c("mu", "sigma", "nu", "tau")}

\item{...}{Other parameters}
}
\value{
A dataset with the estimated random effects. The first column "sa" denotes the Samll Area
}
\description{
Extract random effects for object of class "saegamlss"
}
\note{
The function ranef is firstly imported from nlme and than adapted for object of class "saegamlss"
}
\examples{
index_est <- sa_p_index(sample = s_data, y = "y",
                        sa = "sa", fdis = "LOGNO",
                        sigma.f = TRUE, index = "all")

ranef(object = index_est, parameters = c("mu", "sigma"))


}
