% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_this.R
\name{replace_term}
\alias{replace_term}
\title{Replace term in functions}
\usage{
replace_term(f1, old, new)
}
\arguments{
\item{f1}{a function}

\item{old}{an old term to be replaced}

\item{new}{the new term}
}
\value{
a function
}
\description{
Replace term in functions
}
