% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_this.R
\name{rmNullObs}
\alias{rmNullObs}
\title{Recursively step down into list, removing all such objects}
\usage{
rmNullObs(x)
}
\arguments{
\item{x}{a list}
}
\value{
a list
}
\description{
Recursively step down into list, removing all such objects
}
