% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_data.R
\docType{data}
\name{s_data}
\alias{s_data}
\title{Sample data}
\format{
An object of class \code{"data.frame"}

\describe{
 \item{y}{The dependent variable. The total expediture}
 \item{x1}{A continous covariate}
 \item{x2}{A continous covariate}
 \item{x3}{A continous covariate}
 \item{sa}{The Small Areas}
 \item{id}{The Id of the units}
 \item{pi}{Inclusion probability of the first order}
 \item{s_weights}{Sample weights}
 \item{ncomp}{Number of components of the household}
}
}
\source{
{Created by authors following the Household Budget Survey structure}
}
\usage{
data(s_data)
}
\description{
A data set of a simulated sample survey. Contains the dependent variable (y), the covariates and design information
}
\keyword{datasets}
