% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_p_index.R
\name{sa_p_index}
\alias{sa_p_index}
\title{Estimation of poverty indicators with simplified SAE-GAMLSS}
\usage{
sa_p_index(
  data,
  y,
  sigma.f = TRUE,
  nu.f = TRUE,
  tau.f = TRUE,
  sa,
  w = NULL,
  fdis,
  index = "all",
  epsilon = 1,
  seed = 123
)
}
\arguments{
\item{data}{The dataset with sampled data}

\item{y}{The dependent variable}

\item{sigma.f}{Logical value if TRUE (default) a random effect is used for sigma}

\item{nu.f}{Logical value if TRUE (default) a random effect is used for nu}

\item{tau.f}{Logical value if TRUE (default) a random effect is used for tau}

\item{sa}{The name of the variable to be used as "small area"}

\item{w}{Sample weights}

\item{fdis}{The assumed distribution. Options are: GB2 (Generalized Beta of 2-type), GAMMA (Gamma), EXP (Exponential), LOGNO (Log-Normal), PARETO (Pareto), WEI (Weibull)}

\item{index}{The index to be estimated ("Gini", "Theil" or "Atkinson"). Default is all}

\item{epsilon}{The value for the poverty aversion parameter. Default value is set to 1}

\item{seed}{The seed. Default is 123}
}
\value{
An object of class "saegamlss_class" with the estimated indicators and the value of
the estimated distribution parameters for each area
}
\description{
Estimate the values of poverty indicators in small areas using the simplified SAE-GAMLSS.
}
\examples{

##################
###Using s_data###
##################


index_est <- sa_p_index(data = s_data, y = "y",
                        sa = "sa", fdis = "LOGNO",
                        sigma.f = TRUE, index = "all")

index_est$estimates_par
index_est$Gini

}
\author{
Lorenzo Mori and Maria Rosaria Ferrante
}
