% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saegamlss-package.R
\docType{package}
\name{saegamlss-package}
\alias{saegamlss}
\alias{saegamlss-package}
\title{saegamlss: Unit-Level small area estimation using Generalized Additive Models for Location, Scale and Shape}
\description{
A set of functions to estimate Small Area Unit Level based on Generalized Additive Models for Location, Scale and Shape, including functions for mean squared error estimation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/llorenxo/saegamlss}
  \item Report bugs at \url{https://github.com/llorenxo/saegamlss/issues}
}

}
\keyword{internal}
