% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theil_Exp.R
\name{theil.exp}
\alias{theil.exp}
\title{Theil index exponential distribution}
\usage{
theil.exp(mu)
}
\arguments{
\item{mu}{The estimated value of mu}
}
\value{
The Theil index
}
\description{
Compute the Theil index with exponential distribution
}
