% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theil_Gamma.R
\name{theil.gamma}
\alias{theil.gamma}
\title{Theil index Gamma distribution}
\usage{
theil.gamma(sigma)
}
\arguments{
\item{sigma}{The estimated value of sigma}
}
\value{
The Theil index
}
\description{
Compute the Theil index with Gamma distribution
}
