% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theil_GB2.R
\name{theil.gb2}
\alias{theil.gb2}
\title{Theil index GB2 distribution}
\usage{
theil.gb2(mu, sigma, nu, tau)
}
\arguments{
\item{mu}{The estimated value of mu}

\item{sigma}{The estimated value of sigma}

\item{nu}{The estimated value of nu}

\item{tau}{The estimated value of tau}
}
\value{
The Theil index
}
\description{
Compute the Theil index with GB2 distribution
}
