% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theil_Logno.R
\name{theil.logno}
\alias{theil.logno}
\title{Theil index Log-Normal distribution}
\usage{
theil.logno(sigma)
}
\arguments{
\item{sigma}{The estimated value of sigma}
}
\value{
The Theil index
}
\description{
Compute the Theil index with Log-Normal distribution
}
