% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theil_Pareto.R
\name{theil.pareto}
\alias{theil.pareto}
\title{Theil index Pareto distribution}
\usage{
theil.pareto(sigma)
}
\arguments{
\item{sigma}{The estimated value of sigma}
}
\value{
The Theil index
}
\description{
Compute the Theil index with Pareto distribution
}
