% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theil_Weibull.R
\name{theil.weibull}
\alias{theil.weibull}
\title{Theil index Weibull distribution}
\usage{
theil.weibull(sigma)
}
\arguments{
\item{sigma}{The estimated value of sigma}
}
\value{
The Theil index
}
\description{
Compute the Theil index with Weibull distribution
}
