% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtable.saegamlss_class.R
\name{xtable.saegamlss}
\alias{xtable.saegamlss}
\title{Create Export Tables}
\usage{
\method{xtable}{saegamlss}(
  x,
  caption = NULL,
  label = NULL,
  align = NULL,
  digits = NULL,
  display = NULL,
  auto = FALSE,
  ...
)
}
\arguments{
\item{x}{An R object of class "saegamlss"}

\item{caption}{Character vector of length 1 or 2 containing the table's caption or title. If length is 2, the second item is the "short caption" used when LaTeX generates a "List of Tables". Set to NULL to suppress the caption. Default value is NULL}

\item{label}{Character vector of length 1 containing the LaTeX label or HTML anchor. Set to NULL to suppress the label. Default value is NULL}

\item{align}{Character vector of length equal to the number of columns of the resulting table, indicating the alignment of the corresponding columns}

\item{digits}{Numeric vector of length equal to one (in which case it will be replicated as necessary) or to the number of columns of the resulting table or matrix of the same size as the resulting table, indicating the number of digits to display in the corresponding columns}

\item{display}{Character vector of length equal to the number of columns of the resulting table, indicating the format for the corresponding columns}

\item{auto}{Logical, indicating whether to apply automatic format when no value is passed to align, digits, or display. This ‘autoformat’ (based on xalign, xdigits, and xdisplay) can be useful to quickly format a typical matrix or data.frame. Default value is FALSE}

\item{...}{Other parameters}
}
\value{
an xtable
}
\description{
Create Export Tables for object of class "saegamlss"
}
\examples{

est <-  sa_p_index(sample = s_data, y = "y",
                   sa = "sa", fdis = "LOGNO",
                   sigma.f = TRUE, index = "all")

xtable(x = est)

}
