% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{data_gen}
\alias{data_gen}
\title{Data generation}
\usage{
data_gen(
  Ni,
  M,
  D,
  k,
  b1,
  x1,
  b2 = NULL,
  x2 = NULL,
  b3 = NULL,
  x3 = NULL,
  b4 = NULL,
  x4 = NULL,
  xh = NULL,
  Dis,
  l,
  sigma = NULL,
  sigmah = NULL,
  sigmae = NULL,
  ty = NULL,
  costh = NULL,
  seed = NULL,
  id = NULL
)
}
\arguments{
\item{Ni}{1xD vector containing the number of units (in population) for each area}

\item{M}{Number of replicates}

\item{D}{Number of areas}

\item{k}{A vector, of maximum length equal to 4, of constant (intercept) used to generate data, i.e. mu=k+b1*x}

\item{b1}{A vector of regression coefficients used for mu}

\item{x1}{A matrix or a dataset where each column is a covariate used for mu}

\item{b2}{A vector of regression coefficients used for sigma}

\item{x2}{A matrix or a dataset where each column is a covariate used for sigma}

\item{b3}{A vector of regression coefficients used for nu}

\item{x3}{A matrix or a dataset where each column is a covariate used for nu}

\item{b4}{A vector of regression coefficients used for tau}

\item{x4}{A matrix or a dataset where each column is a covariate used for tau}

\item{xh}{A vector of covariate used to generate heteroschedasticity}

\item{Dis}{Type of distribution in form of rDis where Dis is one of the distribution allowe by GAMLSS. It is to be defined to use distribution different from the nornal one}

\item{l}{A vector,  of maximum length equal to 4, in which each space is a function (the inverse of the link-function) to be applied to the corresponding paramter, i.e. mu, sigma, nu and tau.}

\item{sigma}{A vector,  of maximum length equal to 4, of standard deviation used to generate random effects}

\item{sigmah}{The standard deviation used to generate heteroschedasticity}

\item{sigmae}{The standard deviation used to generate the error term when the link-function is the identity one}

\item{ty}{To be specified equal to "no" if the generation process is mu=k+b1*x+e  or equal to "het" if the generation process have to include heteroschedasticity}

\item{costh}{A costant used to generate heteroschedasticity}

\item{seed}{The seed to be used}

\item{id}{Unit id.}
}
\value{
A list of data frame. Each one has: the dependent variable y, a set of fixed covariates, the small area column (sa) and the id column (id)
}
\description{
A tool to generate the dependent variable, with fixed covariates.

The dependent variable could be generate for more than 100 distributions    which have at maximum 4 parameters. Each parameter could be defined in terms of covariates and random-effects.

The dependent variable, when normal, could include both an additional error-term and heteroschedasticity.
}
\note{
The definition of the heteroschedastic term follows Ramirez-Aldana, R., & Naranjo, L. (2021).
}
\examples{
#Normal data
data_gen(Ni=rep(10,4), D=4, M=2 ,ty ="no", k=100, b1=4,
         x1=rnorm(40, 0,1),  b2 = NULL, x2 = NULL, b3 = NULL,
         x3 = NULL, b4 = NULL, x4 = NULL, xh = NULL,
         Dis, l=c(identity), sigma = 6,sigmah = NULL,
         sigmae = 22,  costh = NULL, seed = 1234)
#
#Heteroschedastic Normal data
#
data_gen(Ni=rep(10,4), D=4, M=2 ,ty ="het", k=100, b1=4,
         x1=rnorm(40, 0,1), b2 = NULL, x2 = NULL, b3 = NULL,
         x3 = NULL, b4 = NULL, x4 = NULL, xh = rnorm(40, 0,1),
         Dis, l=c(identity), sigma = 6, sigmah = 1, sigmae = 22,
         costh = 0.1, seed = 1234)
#
#Non-normal data:
#
#a) Log-normal data
#
data_gen(Ni=rep(10,4), D=4, M=2 ,ty =NULL, k=c(7,-2), b1=1,
         x1=rnorm(40, 0,1), b2 = 0.5, x2 = rnorm(40, 0,1),
         b3 = NULL, x3 = NULL, b4 = NULL, x4 = NULL,
         xh = NULL, Dis=rLOGNO2,l=c(exp, exp),
         sigma = c(0.4, 0.3), sigmah = NULL,  sigmae = NULL,
         costh = NULL, seed = 1234)
#
#b) Dagum data
#
data_gen(Ni=rep(10,4), D=4, M=2 ,ty =NULL,
         k=c(3, log(3.4), -0.4, log(1)), b1=1.5,
         x1=rnorm(40, 0,1), b2 = 0, x2 =NULL,
         b3 = 0.1, x3 = rnorm(40, 0,1), b4 = 0, x4 = NULL,
         xh = NULL, Dis=rGB2, l=c(exp, exp, exp ,exp),
         sigma = c(0.15,0, 0, 0), sigmah = NULL, sigmae = NULL,
         costh = NULL, seed = 1234)
}
\references{
Mori, L., & Ferrante, M. R. (2023). Small area estimation under unit-level generalized additive models for location, scale and shape. arXiv e-prints, arXiv-2302.

Ramirez-Aldana, R., & Naranjo, L. (2021). Random intercept and linear mixed models including heteroscedasticity in a logarithmic scale: Correction terms and prediction in the original scale. PloS one, 16(4), e0249910.

Rigby, R. A., & Stasinopoulos, D. M. (2005). Generalized additive models for location, scale and shape. Journal of the Royal Statistical Society: Series C (Applied Statistics), 54(3), 507–554.
}
\author{
Lorenzo Mori and Maria Rosaria Ferrante
}
