% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_HT.R
\name{direct_HT}
\alias{direct_HT}
\title{Direct Horvitz-Thompson estimatorof the mean}
\usage{
direct_HT(
  y,
  sa,
  pi,
  pi2 = NULL,
  N = NULL,
  var_est = TRUE,
  var_method = "bootstrapSRS",
  B = 100,
  fpc = TRUE,
  data
)
}
\arguments{
\item{y}{The name of the responde variable}

\item{sa}{The name of the Small Areas}

\item{pi}{The name of the first order inclusion probability}

\item{pi2}{A square matrix of the joint inclusion probabilities. Needed for the "LinHT" variance estimator}

\item{N}{A numeric value of the population size for each area. If NULL, it is estimated with the sum of the inverse of the pis}

\item{var_est}{A logical indicating whether or not to compute a variance estimator. Default is TRUE}

\item{var_method}{The method to use when computing the variance estimator. Options are a Taylor linearized technique: "LinHB"= Hajek-Berger estimator, "LinHH" = Hansen-Hurwitz estimator, "LinHTSRS" = Horvitz-Thompson estimator under simple random sampling without replacement, and "LinHT" = Horvitz-Thompson estimator or a resampling technique: "bootstrapSRS" = bootstrap variance estimator under simple random sampling without replacement. The default is "bootstrapSRS"}

\item{B}{The number of bootstrap samples if computing the bootstrap variance estimator. Default is 100}

\item{fpc}{Default to TRUE, logical for whether or not the variance calculation should include a finite population correction when calculating the "LinHTSRS" or the "SRSbootstrap" variance estimator}

\item{data}{The data-set containing y, sa, and pi}
}
\value{
A list with the Horvitz-Thompson estimate of the mean, the variance, the Standard Deviation and the Coefficient of Variation for each Area
}
\description{
Compute the the Horvitz-Thompson Estimator for a finite population mean/proportion based on sample data
collected from a complex sampling design for Small Area.
}
\examples{

data = data.frame("y"= c(1:10), "sa"=sample(c("a","b"), 10, replace=TRUE),
                  "pi"=rep(0.1, 10))
direct_HT(y = "y", sa = "sa", pi = "pi", data = data)
}
