% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_saegamlss.R
\name{est_saegamlss}
\alias{est_saegamlss}
\title{Monte Carlo estimation of mean and HCR based on SAE GAMLSS}
\usage{
est_saegamlss(
  sample,
  nonsample,
  y_dip,
  sa,
  f1,
  f2 = NULL,
  f3 = NULL,
  f4 = NULL,
  fdis,
  R = 50,
  Dis,
  param = "both",
  tau.fix = NULL,
  nu.fix = NULL,
  z = NULL,
  seed = 123
)
}
\arguments{
\item{sample}{A dataset of sampled units.}

\item{nonsample}{A dataset of non-sampled units. With covariates and small areas}

\item{y_dip}{The dependent variable}

\item{sa}{The name of the variable containing the areas}

\item{f1}{A formula object for fitting a model for the mu parameter, e.g.f1=y~x+random(sa)}

\item{f2}{A formula object for fitting a model for the sigma parameter, e.g.f2=y~x+random(sa)}

\item{f3}{A formula object for fitting a model for the nu parameter, e.g.f3=y~x+random(sa)}

\item{f4}{A formula object for fitting a model for the tau parameter, e.g.f4=y~x+random(sa)}

\item{fdis}{The distribution family of the GAMLSS object}

\item{R}{Number of Monte-Carlo repetition. Default is 50}

\item{Dis}{Type of distribution in form of rDis where Dis is one of the distribution allowed by GAMLSS}

\item{param}{The parameter to estimate: "mean", "HCR" or "both". Default is both. A new parameter to be estimated can be passed as a function}

\item{tau.fix}{A value to be fixed to 1 to obtain the Singh-Maddala distribution. Dist must be GB2}

\item{nu.fix}{A value to be fixed to 1 to obtain the Dagum distribution. Dist must be GB2}

\item{z}{The Poverty line. Default is equal to 0.6 of the dependent variable}

\item{seed}{The seed. Default is 123}
}
\value{
an object of class "saegamlss_class" which contains two list. The first list, named est has:
 ME 1xD vector with the estimate of the mean and
 HCR 1xD vector with the estimate of the HCR.
 The second, named input_var has:
 y the dependent variable
 sa the name of the variable cotaining the areas
 fit a GAMLSS object with all the components of the regression
 f1 a formula object used for fitting a model to the mu parameter
 f2 a formula object used for fitting a model to the sigma parameter
 f3 a  formula object used for fitting a model to the nu parameter
 f4 a formula object used for fitting a model to the tau parameter
 R the number of loop used to obtain the monte-carlo estimation
 D the number of area
 Ni 1xD vector containing the number of units (in population) for each area
 originaldata The dataset of sampled-units
 fids The distribution used for the estimation
 z The poverty line
}
\description{
A function to estimate the mean or/and the HCR     derived from unit-level small area estimation based on generalized additive models for location, scale and shape
}
\note{
With "object"$input_var$fit is possible to use all the classical function used by gamlss
 The Small Area have to be denoted with a number from 1 to D
}
\examples{

##################
###Using s_data###
##################

est <- est_saegamlss(
  sample = s_data, nonsample = pop_data, y_dip="y",
  sa="sa", f1 = y ~ x1 + random(sa), f2 = y ~ x2 + random(sa),
  f3 = NULL, f4 = NULL, fdis = NO, R = 20,
  Dis = rNO,  param = "mean",
  tau.fix = NULL, nu.fix = NULL
)

est$estimates

# Defining a new parameter to be estimated

est <- est_saegamlss(
  sample = s_data, nonsample = pop_data, y_dip="y",
  sa="sa", f1 = y ~ x1 + random(sa), f2 = y ~ x2 + random(sa),
  f3 = NULL, f4 = NULL, fdis = NO, R = 20,
  Dis = rNO, param = function(x) (mean(x^2)),
  tau.fix = NULL, nu.fix = NULL
 )
est$estimates

}
\references{
Mori, L., & Ferrante, M. R. (2023). Small area estimation under unit-level generalized additive models for location, scale and shape. arXiv e-prints, arXiv-2302.
 Graf, M., Marin, J. M., & Molina, I. (2019). A generalized mixed model for skewed distributions applied to small area estimation. Test, 28(2), 565–597.
}
\author{
Lorenzo Mori and Maria Rosaria Ferrante
}
