% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_selection.R
\name{m_selection}
\alias{m_selection}
\title{Small area model selection within GAMLSS}
\usage{
m_selection(
  sample_data,
  y,
  f_cov,
  kp = 2,
  ndis = 3,
  R = 200,
  k = 7,
  type = NULL,
  fix_dis = NULL,
  supp = TRUE,
  seed = 123
)
}
\arguments{
\item{sample_data}{The dataset with sampled units}

\item{y}{The dependent variable}

\item{f_cov}{A formula containing all the possible covariates and/or additive terms (i.e. x1+x2+x3+random(x4))}

\item{kp}{the penalty for the GAIC (Step 1) with default values kp=2 the standard AIC.}

\item{ndis}{Number of distribution to be consider at the first step. Default is 3}

\item{R}{Number of loop to be done within the k-fold cross validation}

\item{k}{Number of fold in k-fold cross validation. Default is 7}

\item{type}{Type of step 2 to be done. A or B. Default is B (see references)}

\item{fix_dis}{A distribution to be tested even if is not selected within the ndis at step 1}

\item{supp}{Suppress the warning in the three step. Default is TRUE}

\item{seed}{The seed. Default is 123}
}
\value{
An object of class "saegamlss_class", named results, with the results of each steps and the GAIC of all the tested distributions at step 1.
}
\description{
The function select the final SAE-GAMLSS model following three steps as described by Mori and Ferrante (2023)
}
\note{
The summary (Step2) do not reports results on the random effects as usual for GAMLSS. See Step2 select_v for the random effects
}
\examples{

sample_data=data_gen(
  Ni = rep(10, 4), D = 4, M = 1, ty = "no", k = 100, b1 = 4,
  x1 = rnorm(40, 0, 1), b2 = NULL, x2 = NULL, b3 = NULL,
  x3 = NULL, b4 = NULL, x4 = NULL, xh = NULL,
  Dis = rNO, l = c(identity), sigma = 6, sigmah = NULL,
  sigmae = 22, costh = NULL
)

#Adding x2
set.seed(1234)
sample_data=as.data.frame(sample_data[[1]])
sample_data$x2=rnorm(40, 0, 1)

m_selection(sample_data=sample_data, y = sample_data$y,
            f_cov= ~x1+x2+random(sa), ndis=1,
            R=2, k=1, type="A", fix_dis="NO")
}
\references{
Mori, L., & Ferrante, M. R. (2023). Small area estimation under unit-level generalized additive models for location, scale and shape. arXiv e-prints, arXiv-2302.
 Stasinopoulus et al. (2019). Flexible regression and smoothing using GAMLSS in R.
 Rigby, R. A., & Stasinopoulos, D. M. (2005). Generalized additive models for location, scale and shape. Journal of the Royal Statistical Society: Series C (Applied Statistics), 54(3), 507–554.
}
\author{
Lorenzo Mori and Maria Rosaria Ferrante
}
