% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse_saegamlss.R
\name{mse_saegamlss}
\alias{mse_saegamlss}
\title{Bootstrap Mean Square Error for SAE GAMLSS}
\usage{
mse_saegamlss(
  est,
  loop = 200,
  l,
  Iden = FALSE,
  cov1,
  cov2 = NULL,
  cov3 = NULL,
  cov4 = NULL,
  seed = 123
)
}
\arguments{
\item{est}{An object obtained with est_saegamlss()}

\item{loop}{Number of loop of bootstrap. Default is 200}

\item{l}{A vector, of maximum length equal to 4, in which each space is a function (the inverse of the link-function) to be applied to the corresponding parameter, i.e. mu, sigma, nu and tau}

\item{Iden}{TRUE if the Identity link function is used for mu. Default is FALSE}

\item{cov1}{A matrix or a data frame with covariates for the whole population used for mu. If an intercept is used the first columns have to be a vector of 1}

\item{cov2}{A matrix or a data frame with covariates for the whole population used for sigma. If an intercept is used the first columns have to be a vector of 1}

\item{cov3}{A matrix or a data frame with covariates for the whole population used for nu. If an intercept is used the first columns have to be a vector of 1}

\item{cov4}{A matrix or a data frame with covariates for the whole population used for tau. If an intercept is used the first columns have to be a vector of 1}

\item{seed}{The seed. Default is 123}
}
\value{
an object of class "saegamlss" containing the MSE of the estimated parameter for each area
        and all the values returned by est_saegamlss()
}
\description{
Compute the parametric bootstrap MSE for the estimation of the mean and the HCR or a self-defined parameter.
}
\examples{

##################
###Using s_data###
##################


est <- est_saegamlss(
  sample = s_data, nonsample = pop_data, y_dip="y",
  sa="sa", f1 = y ~ x1 + random(sa), f2 = y ~ x2 + random(sa),
  f3 = NULL, f4 = NULL, fdis = NO, R = 2,
  Dis = rNO,  param = "mean",
  tau.fix = NULL, nu.fix = NULL
)

data <- rbind(est$input_var$origindata  \%>\%
              dplyr::select( dplyr::all_of(colnames(est$input_var$nonsample))),
              est$input_var$nonsample
               )

x <- data.frame(rep(1, nrow(data)), "x1" = data$x1,
                "x2" = data$x2)


MSE <- mse_saegamlss(
  est = est, loop = 2,
  l = c(identity, exp), Iden = TRUE,
  cov1 = x[,1:2], cov2 = x[,c(1,3)], cov3 = NULL,
  cov4 = NULL, seed = 124
)

MSE$est_mse


#MSE for a self-defined parameter

est <- est_saegamlss(
  sample = s_data, nonsample = pop_data, y_dip = "y",
  sa = "sa", f1 = y ~ x1 + random(sa), f2 = y ~ x2 + random(sa),
  f3 = NULL, f4 = NULL, fdis = NO, R = 2,
  Dis = rNO, param = function(x) (mean(x^2)),
  tau.fix = NULL, nu.fix = NULL
 )


MSE <- mse_saegamlss(
  est = est, loop = 2,
  l = c(identity, exp), Iden = TRUE,
  cov1 = x[,1:2], cov2 = x[,c(1,3)], cov3 = NULL,
  cov4 = NULL
)


MSE$est_mse

}
\references{
Mori, L., & Ferrante, M. R. (2024). Small area estimation under unit-level generalized additive models for location, scale and shape. Journal of Survey Statistics and Methodology, 1–37
 Graf, M., Marin, J. M., & Molina, I. (2019). A generalized mixed model for skewed distributions applied to small area estimation. Test, 28(2), 565–597
}
\author{
Lorenzo Mori and Maria Rosaria Ferrante
}
