% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonsample.R
\name{nonsample}
\alias{nonsample}
\title{Create the non-sample dataset}
\usage{
nonsample(data, sample, id)
}
\arguments{
\item{data}{Population dataset}

\item{sample}{Sample dataset}

\item{id}{id variable, must be the same in both the datasets}
}
\value{
A dataset of non-sampled units
}
\description{
Subset the non-sample part of the population
}
