% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_mse_ssaegamlss.R
\name{np_mse_ssaegamlss}
\alias{np_mse_ssaegamlss}
\title{Non-parametric MSE for simplified SAE-GAMLSS}
\usage{
np_mse_ssaegamlss(
  data,
  y,
  sa,
  ncomp,
  R = 200,
  sigma.f = TRUE,
  nu.f = TRUE,
  tau.f = TRUE,
  w = NULL,
  fdis,
  nRS = 20,
  nCG = 20,
  index = NULL,
  epsilon = NULL
)
}
\arguments{
\item{data}{A dataset containing household values}

\item{y}{The dependent variable}

\item{sa}{The Small Area domains}

\item{ncomp}{The number of components of each household}

\item{R}{The number of loops to be performed. Default is 200}

\item{sigma.f}{Logical value if TRUE (default) a random effect is used for sigma}

\item{nu.f}{Logical value if TRUE (default) a random effect is used for nu}

\item{tau.f}{Logical value if TRUE (default) a random effect is used for tau}

\item{w}{Sample weights}

\item{fdis}{The assumed distribution. Options are: GB2 (Generalized Beta of 2-type), GA (Gamma), EX (Exponential), LOGNO (Log-Normal), PA (Pareto), WE (Weibull)}

\item{nRS}{Number of loop to do with the RS() algorithm}

\item{nCG}{Number of loop to do with the CG() algorithm}

\item{index}{The index to be estimated ("Gini", "Theil" or "Atkinson"). Default is all}

\item{epsilon}{The value for the poverty aversion parameter. Default value is set to 1}
}
\value{
An object of class "saegamlss_class" containing the values of the MSE for each area and for each index
}
\description{
Compute the Non-parametric MSE for simplified SAE-GAMLSS for three possible different
indicators (Gini, Theil, Atkinson)
}
\examples{

set.seed(124)

data <- data.frame("y"=rLOGNO(20, mu=8, sigma=0.8),
"sa" =  as.factor(rep(c(1,2,3,4,5),4)),
"ncomp"=rep(c(2,1,3,4), 5),
"w"=1/runif(20, 0.5, 1))

<<<<<<< HEAD
np<- np_mse_ssaegamlss(data=data, y=data$y, sa = data$sa,
                       ncomp=data$ncomp, fdis="LOGNO",
                       index="Gini",
                       nRS=2, nCG=2, R=2)
=======
np<- np_mse_ssaegamlss(data=data, y=data$y, sa = data$sa, ncomp=data$ncomp, fdis="LOGNO", index="Gini",
                  nRS=2, nCG=2, R=2)
>>>>>>> a74883390a3cda411eefbe008ee4d6279fd79702

np

}
\author{
Lorenzo Mori and Maria Rosaria Ferrante
}
