% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_2.R
\docType{data}
\name{pi_2}
\alias{pi_2}
\title{Second Order Inclusion Probability Data}
\format{
An object of class \code{"data.frame"} with the following columns:
}
\source{
{Created by authors following the Household Budget Survey structure}
}
\usage{
data(pi_2)
}
\description{
A data set containing second order inclusion probabilities. The main diagonal holds the first order inclusion probabilities. This dataset can be used, for example, in the function \code{direct_HT()}.
}
\keyword{datasets}
