% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saegamlss-package.R
\docType{package}
\name{saegamlss-package}
\alias{saegamlss}
\alias{saegamlss-package}
\title{saegamlss: Unit-Level small area estimation using Generalized Additive Models for Location, Scale and Shape}
\description{
This package provides a set of functions for Small Area Unit Level based on Generalized Additive Models for Location, Scale and Shape, including functions for mean squared error estimation.
}
\keyword{internal}
